refactopr: move d3functions to new directory
