+++ javascript/src/viz/helpers/d3functions/curves.js
